build: decrease cfg(unix) gates
