+++ core/Cargo.toml
@@ -77,9 +77,11 @@ rand = { workspace = true }
 rand_chacha = { workspace = true }
 rayon = { workspace = true }
 rolling-file = { workspace = true }
+rts-alloc = { workspace = true }
 rustls = { workspace = true }
 serde = { workspace = true }
 serde_bytes = { workspace = true }
+shaq = { workspace = true }
 signal-hook = { workspace = true }
 slab = { workspace = true }
 solana-account = { workspace = true }
@@ -171,8 +173,6 @@ trees = { workspace = true }
 wincode = { workspace = true }
 
 [target."cfg(unix)".dependencies]
-rts-alloc = { workspace = true }
-shaq = { workspace = true }
 sysctl = { workspace = true }
 
 [dev-dependencies]
