+++ core/src/banking_stage/tpu_to_pack.rs
@@ -4,7 +4,7 @@
 use {
     agave_banking_stage_ingress_types::BankingPacketReceiver,
     agave_scheduler_bindings::{SharableTransactionRegion, TpuToPackMessage, tpu_message_flags},
-    agave_scheduling_utils::handshake::server::AgaveTpuToPackSession,
+    agave_scheduling_utils::handshake::AgaveTpuToPackSession,
     rts_alloc::Allocator,
     solana_packet::PacketFlags,
     solana_perf::packet::PacketBatch,
