+++ scheduling-utils/src/handshake/mod.rs
@@ -1,7 +1,9 @@
+#[cfg(unix)]
 pub mod client;
+#[cfg(unix)]
 pub mod server;
 mod shared;
 #[cfg(test)]
 mod tests;
 
-pub use shared::{ClientLogon, MAX_WORKERS, logon_flags};
+pub use shared::*;
