+++ scheduling-utils/src/handshake/shared.rs
@@ -1,3 +1,14 @@
+use {
+    agave_scheduler_bindings::{
+        PackToWorkerMessage, ProgressMessage, TpuToPackMessage, WorkerToPackMessage,
+    },
+    rts_alloc::Allocator,
+    thiserror::Error,
+};
+
+pub(crate) type RtsAllocError = rts_alloc::error::Error;
+pub(crate) type ShaqError = shaq::error::Error;
+
 pub const MAX_WORKERS: usize = 64;
 
 /// Protocol version.
@@ -49,3 +60,80 @@ impl ClientLogon {
 }
 
 pub mod logon_flags {}
+
+/// The complete initialized scheduling session.
+pub struct ClientSession {
+    pub allocators: Vec<Allocator>,
+    pub tpu_to_pack: shaq::spsc::Consumer<TpuToPackMessage>,
+    pub progress_tracker: shaq::spsc::Consumer<ProgressMessage>,
+    pub workers: Vec<ClientWorkerSession>,
+}
+
+/// An per worker scheduling session.
+pub struct ClientWorkerSession {
+    pub pack_to_worker: shaq::spsc::Producer<PackToWorkerMessage>,
+    pub worker_to_pack: shaq::spsc::Consumer<WorkerToPackMessage>,
+}
+
+/// Potential errors that can occur during the client's side of the handshake.
+#[derive(Debug, Error)]
+pub enum ClientHandshakeError {
+    #[error("Io; err={0}")]
+    Io(#[from] std::io::Error),
+    #[error("Timed out")]
+    TimedOut,
+    #[error("Protocol violation")]
+    ProtocolViolation,
+    #[error("Rejected; reason={0}")]
+    Rejected(String),
+    #[error("Rts alloc; err={0}")]
+    RtsAlloc(#[from] RtsAllocError),
+    #[error("Shaq; err={0}")]
+    Shaq(#[from] ShaqError),
+}
+
+/// An initialized scheduling session.
+pub struct AgaveSession {
+    pub flags: u16,
+    pub tpu_to_pack: AgaveTpuToPackSession,
+    pub progress_tracker: shaq::spsc::Producer<ProgressMessage>,
+    pub workers: Vec<AgaveWorkerSession>,
+}
+
+/// Shared memory objects for the tpu to pack worker.
+pub struct AgaveTpuToPackSession {
+    pub allocator: Allocator,
+    pub producer: shaq::spsc::Producer<TpuToPackMessage>,
+}
+
+/// Shared memory objects for a single banking worker.
+pub struct AgaveWorkerSession {
+    pub allocator: Allocator,
+    pub pack_to_worker: shaq::spsc::Consumer<PackToWorkerMessage>,
+    pub worker_to_pack: shaq::spsc::Producer<WorkerToPackMessage>,
+}
+
+/// Potential errors that can occur during the Agave side of the handshake.
+///
+/// # Note
+///
+/// These errors are stringified (up to 256 bytes then truncated) and sent to the client.
+#[derive(Debug, Error)]
+pub enum AgaveHandshakeError {
+    #[error("Io; err={0}")]
+    Io(#[from] std::io::Error),
+    #[error("Timeout")]
+    Timeout,
+    #[error("Close during handshake")]
+    EofDuringHandshake,
+    #[error("Version; server={server}; client={client}")]
+    Version { server: u64, client: u64 },
+    #[error("Worker count; count={0}")]
+    WorkerCount(usize),
+    #[error("Allocator handles; count={0}")]
+    AllocatorHandles(usize),
+    #[error("Rts alloc; err={0:?}")]
+    RtsAlloc(#[from] RtsAllocError),
+    #[error("Shaq; err={0:?}")]
+    Shaq(#[from] ShaqError),
+}
