+++ scheduling-utils/src/lib.rs
@@ -3,13 +3,8 @@
 pub mod error;
 pub mod thread_aware_account_locks;
 
-#[cfg(unix)]
 pub mod bridge;
-#[cfg(unix)]
 pub mod handshake;
-#[cfg(unix)]
 pub mod pubkeys_ptr;
-#[cfg(unix)]
 pub mod responses_region;
-#[cfg(unix)]
 pub mod transaction_ptr;
